View 연동 완료 ..!
