+++ ParcelV2Widget/ParcelV2Widget.swift
@@ -62,23 +62,9 @@ struct ParcelV2WidgetEntryView : View {
     
     var body: some View {
         if let parcel = entry.parcel {
-            VStack {
-                if let date = dateToString(entry.date) {
-                    HStack {
-                        Text("업데이트시간: \(date)")
-                            .font(.system(size: 8))
-                        Spacer()
-                    }
-                }
-                
-                Spacer()
-                if let item = parcel.item {
-                    Text(item)
-                }
-                Text(parcel.state.text)
-                Spacer()
-            }
-            .padding([.all], 16)
+            DeliveryView(parcel: parcel)
+        } else {
+            Text("no data")
         }
     }
     
@@ -86,25 +72,6 @@ struct ParcelV2WidgetEntryView : View {
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "HH:mm"
         return dateFormatter.string(from: date)
-    }
-    
-    
-    func convert(type: String)  -> DeliveryDTO? {
-        if "information_received" == type {
-            return .ready
-        } else if "at_pickup" == type || "in_transit" == type {
-            return .moving
-        } else if "out_for_delivery" == type {
-            return .deliveringStart
-        } else if "delivered" == type {
-            return .delivered
-        }
-        
-        return nil
-    }
-    
-    func convertToWrapper(_ data: [Parcel.Progress]) -> [ProgressWrapper] {
-        return data.map { ProgressWrapper(time: $0.time, status: $0.status, location: $0.location, description: $0.description)}
     }
 }
 
@@ -117,52 +84,5 @@ struct ParcelV2Widget: Widget {
         }
         .configurationDisplayName("My Widget")
         .description("This is an example widget.")
-    }
-}
-
-enum DeliveryDTO: String {
-    case ready// information_received //상품준비중
-    case moving// at_pickup, in_transit //상품인수, 상품이동 중 → 이동중
-    case deliveringStart// out_for_delivery //배송출발
-    case delivered// delivered //배송완료
-
-    
-    var stateID: String {
-        switch self {
-        case .ready:
-            return "information_received"
-        case .moving:
-            return "at_pickup"  // "in_transit"
-        case .deliveringStart:
-            return "out_for_delivery"
-        case .delivered:
-            return "delivered"
-        }
-    }
-    
-    var imageString: String {
-        switch self {
-        case .ready:
-            return "shippingbox"
-        case .moving:
-            return "box.truck"
-        case .deliveringStart:
-            return "box.truck"
-        case .delivered:
-            return "house"
-        }
-    }
-    
-    var text: String {
-        switch self {
-        case .ready:
-            return "상품준비중"
-        case .moving:
-            return "이동중"
-        case .deliveringStart:
-            return "배송출발"
-        case .delivered:
-            return "배송완료"
-        }
     }
 }
